requirements.txt
# Core dependencies for Stable Diffusion with DirectML
torch-directml>=0.2.0.dev230426
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0

# Web UI
gradio>=4.12.0

# Image processing
pillow>=10.1.0
numpy>=1.24.0

# Utilities
tqdm>=4.66.0
safetensors>=0.4.1

# Optional but recommended
omegaconf>=2.3.0
